// src/types/types.ts
export interface Category {
  id: number;
  title: string;
  name?: string; // Name if 'title' is unavailable
  slug: string;
  image: string;
}

export interface FetchCategoriesPayload {
  results: Category[];
}

export interface CategoryState {
  categories: Category[];
  status: "idle" | "loading" | "succeeded" | "failed";
  error: string | null;
}

export interface User {
  id: number;
  username: string;
  email: string;
}
export interface Store {
  id: number;
  name: string;
  rating: number;
  reviews_count: number;
  location: string;
  overview: string;
  image: string;
  phone_number?: string;
  whats_app?: string;
  working_hours: string | null;
  is_verified: boolean;
  is_responsive: boolean;
  event_planning_categories: string[];
  rent_hire_categories: string[];
}
// ...
export interface EventCategory {
  id: number;
  title: string;
  slug: string;
  image: string;
  categories?: number[]; // An array of category IDs
}

// EventSubcategory interface to represent event subcategories (e.g., "Pre Wedding Planning")
export interface EventSubcategory {
  id: number;
  title: string;
  slug: string;
  categories: EventCategory[]; // Array of categories within this subcategory
}
// ...
export interface RentHireState {
  rent_hire_categories: Category[];
  stores: Store[];
  rent_hire_subcategories: EventSubcategory[];
  rent_hire_subcategory_details: EventSubcategory | null; // Add this field for specific subcategory details
  status: "idle" | "loading" | "succeeded" | "failed";
  error: string | null;
}